update styles, rename to MoreOptionsToggle
